Move the `doc-common` module under the `@bayou` namespace.
