+++ local-modules/@bayou/doc-common/CaretChange.js
