+++ local-modules/@bayou/doc-common/CaretDelta.js
