+++ local-modules/@bayou/doc-common/CaretSnapshot.js
