+++ local-modules/@bayou/doc-common/Property.js
