+++ local-modules/@bayou/doc-common/PropertyChange.js
