+++ local-modules/@bayou/doc-common/PropertyDelta.js
