+++ local-modules/@bayou/doc-common/PropertySnapshot.js
