+++ local-modules/@bayou/doc-common/TheModule.js
