+++ local-modules/app-common/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "api-common": "local",
     "@bayou/codec": "local",
-    "doc-common": "local",
+    "@bayou/doc-common": "local",
     "@bayou/ot-common": "local",
     "@bayou/util-common": "local"
   }
