+++ local-modules/app-setup/DebugTools.js
@@ -7,7 +7,7 @@ import { camelCase } from 'lodash';
 import { inspect } from 'util';
 
 import { TheModule as appCommon_TheModule } from 'app-common';
-import { DocumentId } from 'doc-common';
+import { DocumentId } from '@bayou/doc-common';
 import { DocServer } from 'doc-server';
 import { AuthorId } from '@bayou/ot-common';
 import { Logger } from '@bayou/see-all';
