+++ local-modules/doc-client/BodyClient.js
@@ -3,7 +3,7 @@
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
 import { ConnectionError } from 'api-common';
-import { BodyChange, BodyDelta, BodySnapshot } from 'doc-common';
+import { BodyChange, BodyDelta, BodySnapshot } from '@bayou/doc-common';
 import { Delay } from '@bayou/promise-util';
 import { QuillEvents, QuillUtil } from 'quill-util';
 import { TString } from '@bayou/typecheck';
