+++ local-modules/doc-client/CaretOverlay.js
@@ -3,7 +3,7 @@
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
 import { DragState } from '@bayou/data-model-client';
-import { CaretOp, CaretSnapshot } from 'doc-common';
+import { CaretOp, CaretSnapshot } from '@bayou/doc-common';
 import { Delay } from '@bayou/promise-util';
 import { QuillEvents, QuillGeometry, QuillUtil } from 'quill-util';
 import { TObject } from '@bayou/typecheck';
