+++ local-modules/doc-client/CaretState.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { CaretSnapshot } from 'doc-common';
+import { CaretSnapshot } from '@bayou/doc-common';
 import { Delay } from '@bayou/promise-util';
 
 /**
