+++ local-modules/doc-client/PropertyClient.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { PropertyDelta, PropertyOp, Timeouts } from 'doc-common';
+import { PropertyDelta, PropertyOp, Timeouts } from '@bayou/doc-common';
 import { TString } from '@bayou/typecheck';
 import { CommonBase, DataUtil, Errors } from '@bayou/util-common';
