+++ local-modules/doc-server/BaseControl.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { Timeouts } from 'doc-common';
+import { Timeouts } from '@bayou/doc-common';
 import { Errors as fileStoreOt_Errors, StoragePath, TransactionSpec } from '@bayou/file-store-ot';
 import { BaseSnapshot, RevisionNumber } from '@bayou/ot-common';
 import { Delay } from '@bayou/promise-util';
