+++ local-modules/quill-util/package.json
@@ -4,7 +4,7 @@
 
   "dependencies": {
     "@bayou/deps-quill-client": "local",
-    "doc-common": "local",
+    "@bayou/doc-common": "local",
     "@bayou/promise-util": "local",
     "@bayou/typecheck": "local",
     "@bayou/util-common": "local"
